cart change to long : - original_amount, - initial_price

+++ features/transaction/cart/src/main/java/com/tokopedia/cart/data/model/response/shopgroupsimplified/PriceChanges.kt
@@ -13,7 +13,7 @@ data class PriceChanges(
 	val changesState: Int = 0,
 
 	@field:SerializedName("original_amount")
-	val originalAmount: Int = 0,
+	val originalAmount: Long = 0,
 
 	@field:SerializedName("description")
 	val description: String = ""
